+++ src/modules/tdk.js
@@ -1,8 +1,23 @@
+var SOZLUK_autoComp;
+
 commands["tdk"] = {
     func: async function (process, isInput = false) {
 
         // if no arguments
         if (process._ == "") {
+
+            if (!SOZLUK_autoComp) {
+                const req = await request("./database/autocomp.json");
+                if (!req) {
+                    stdout.error("Request failed.");
+                    return 500;
+                }
+                SOZLUK_autoComp = await req.json();
+                commands.tdk.autoComplete = SOZLUK_autoComp;
+            }
+    
+            if (!commands.tdk.autoComplete) 
+                commands.tdk.autoComplete = SOZLUK_autoComp;
 
             // set std input prefix to empty
             return ">&nbsp;";
@@ -10,10 +25,10 @@ commands["tdk"] = {
 
         // else make request
         const url = "https://sozluk.gov.tr/gts?ara=" + encodeURI(process._);
-        const req = await request(url, {});
+        const req = await request(url);
         if (!req) {
             stdout.error("Request failed.");
-            return 503;
+            return 500;
         }
         const data = await req.json();
 
@@ -28,7 +43,7 @@ commands["tdk"] = {
         // exit
         return 0;
     },
-    about: `TDK dictionary api.%ALIASES%\nExamples:\n $ tdk merhaba`
+    about: `TDK dictionary api. %ALIASES%\nExamples:\n $ tdk merhaba`
 };
 
 autoCompList.push("tdk");